게시글 상세조회 기능 추가 view 변수에 조회수 기능 추가 반환

+++ src/main/java/com/gamecrew/gamecrew_project/domain/post/service/PostService.java
@@ -4,6 +4,7 @@ import com.gamecrew.gamecrew_project.domain.post.dto.request.PostRequestDto;
 import com.gamecrew.gamecrew_project.domain.post.dto.response.PostResponseDto;
 import com.gamecrew.gamecrew_project.domain.post.entity.Post;
 import com.gamecrew.gamecrew_project.domain.post.repository.PostRepository;
+import com.gamecrew.gamecrew_project.domain.user.entity.User;
 import com.gamecrew.gamecrew_project.global.exception.CustomException;
 import com.gamecrew.gamecrew_project.global.exception.constant.ErrorMessage;
 import com.gamecrew.gamecrew_project.global.response.MessageResponseDto;
@@ -13,30 +14,57 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+
 @Service
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
 
 
-    public void createPost(PostRequestDto requestDto) {
+    public void createPost(PostRequestDto requestDto, User user) {
         // RequestDto -> Entity
-        Post post = new Post(requestDto);
+        Post post = new Post(requestDto, user);
         //DB저장
         Post savePost = postRepository.save(post);
     }
     @Transactional
-    public void updatePost(Long postId, PostRequestDto requestDto) {
+    public void updatePost(Long postId, PostRequestDto requestDto, User user) {
         //해당 메모가 DB에 존재하는지 확인
         Post post = postRepository.findById(postId).orElseThrow(()->
                 new IllegalArgumentException("선택한 글은 존재하지 않습니다."));
+        if (!(post.getUser().getEmail().equals(user.getEmail()))){
+            throw new IllegalArgumentException("작성자만 수정 할 수 있습니다.");}
         // DB수정
         post.update(requestDto);
     }
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, User user) {
         //해당 메모가 DB에 존재하는지 확인
         Post post = postRepository.findById(postId).orElseThrow(()->
                 new IllegalArgumentException("선택한 글은 존재하지 않습니다."));
+        if (!(post.getUser().getEmail().equals(user.getEmail()))){
+            throw new IllegalArgumentException("작성자만 삭제 할 수 있습니다.");}
+
         postRepository.delete(post);
+    }
+    @Transactional
+    public void updateView(Long postId) {
+        Post post = postRepository.findByPostId(postId);
+        post.update();
+    }
+
+    public PostResponseDto getPost(Long postId, User user) {
+        //해당 메모가 DB에 존재하는지 확인
+        Post post = postRepository.findById(postId).orElseThrow(()->
+                new IllegalArgumentException("선택한 글은 존재하지 않습니다."));
+        PostResponseDto postResponseDto;
+        if(post.getUser().getEmail().equals(user.getEmail())){
+            postResponseDto = new PostResponseDto(post);
+            postResponseDto.checkOwner();
+            return postResponseDto;
+        } else {
+            postResponseDto = new PostResponseDto(post);
+            return postResponseDto;
+        }
     }
 }
